+++ runtests.py
@@ -215,9 +215,17 @@ def def_to_cdef(source):
 
     return '\n'.join(output)
 
+
+def exclude_extension_in_pyver(*versions):
+    def check(ext):
+        return EXCLUDE_EXT if sys.version_info[:2] in versions else ext
+    return check
+
+
 def update_linetrace_extension(ext):
     ext.define_macros.append(('CYTHON_TRACE', 1))
     return ext
+
 
 def update_numpy_extension(ext):
     import numpy
@@ -339,6 +347,7 @@ EXT_EXTRAS = {
     'tag:openmp': update_openmp_extension,
     'tag:cpp11': update_cpp11_extension,
     'tag:trace' : update_linetrace_extension,
+    'tag:bytesformat':  exclude_extension_in_pyver((3, 3), (3, 4)),  # no %-bytes formatting
 }